done 5/12

+++ AUSHOP/src/main/java/AUSHOP/repository/KhachHangRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import AUSHOP.entity.KhachHang;
 
-public interface KhachHangRepository extends JpaRepository<KhachHang,Integer> {
+public interface KhachHangRepository extends JpaRepository<KhachHang,Long> {
 
 	Optional<KhachHang> findByEmail(String name);
